Override kafkatee's default logrotate/rsyslog configuration.
This is a preparation step to remove logrotate/syslog configurations
from the deb package. This is currently causing cronspam and logfiles
not being rotated due to the absence of the 'syslog' user on oxygen
(specified explicitly in the logrotate conf instead of root).
The rsyslog/logrotate configurations will be gated by a boolean option
and will leverage the rsyslog/logrotate modules.

Bug: T132322
Change-Id: Ibbc2b6936769a0002729fbd721a89a85b8553206